feat(review): remove utxo_wrapper_block_header_storage.rs
